requirements.txt
alembic
psycopg2
flask
flask-cors
flask-socketio
redis
sqlalchemy-stubs
python-redis-lock
pytest